+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [v0.5.0-alpha] - 2024-09-24
+
+### Features:
+
+- Added items to the game
+- Added a loot table for enemies
+- Added basic item functionality for the player/enemy
+- Added basic consumable items
+- Created an item factory which handles the different types of item creation
+
+### Improvements:
+
+- Expanded existing csv's with more columns and entries
+- Refined existing code
+
+### Bug Fixes:
+
+- NA
+
+### Known bugs:
+
+- The exit game function does not work on the first try when returning to the main menu from the game menu.
+
 ## [v0.4.0-alpha] - 2024-09-23
 
 ### Features:
@@ -16,4 +39,65 @@
 - NA
 
 ### Known bugs:
+- NA
+
+## [v0.3.0-alpha] - 2024-09-22
+
+### Features:
+
+- Added skills
+- Added skill factory
+- Added basic tutorial
+- Incorporated skills into existing classes
+
+### Improvements:
+
+- Refined general code
+
+### Bug Fixes:
+
+- NA
+
+### Known bugs:
+
+- NA
+
+## [v0.2.0-alpha] - 2024-09-20
+
+### Features:
+
+- Added Main menu
+- Added Game menu
+
+### Improvements:
+
+- Refined general code
+
+### Bug Fixes:
+
+- NA
+
+### Known bugs:
+
+- NA
+
+## [v0.1.0-alpha] - 2024-09-19
+
+### Features:
+
+- Added Player
+- Added Enemy
+- Added Battle
+- Very basic functionality
+
+### Improvements:
+
+- NA
+
+### Bug Fixes:
+
+- NA
+
+### Known bugs:
+
 - NA